requirements.txt
streamlit-1.37.1-py2.py3-none-any.whl